Updated mediawiki/extensions Project: mediawiki/extensions/BlogPage  c49428cc11014da4811ec8db0269150822d1d240
Use the page ID in the memcached key to ensure the accuracy of cached data

Special:ArticlesHome by default displays new blog articles on the right
side of the page (call to ArticlesHome::displayNewestPages()). This method
in question calls BlogPage::getVotesForPage() for each returned page to
get the correct amount of votes.
Based on some quick testing on my Labs instance, the SQL query used by
BlogPage::getVotesForPage() does return the correct result, so I'm
guessing this is a cache thing.

Bug: T126109
Change-Id: Icf143fe5e762beadcbe752aa8f1d8a86a8b1375f