Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  1fcba848c577c3dc890a166b3bae228dfd332264
Allowing triggering user tests from query parameter

For the textcat test we want to setup some fairly complex configuration
based on the user being in the test. Providing it all over the query
string would be a bit much, so add the ability for the UserTesting class
to set that configuration based on a query string provided by the
frontend search satisfaction code.

Bug: T121542
Change-Id: I72bc9ae6e09d91dfdc8fbaed6f2704e98a3473c9