Updated mediawiki/extensions Project: mediawiki/extensions/Translate  9bcd655003bfed06022a6794fc52175610bb96d9
Add translate styles via addModuleStyles

Given ext.uls.buttons is a style only module it can
be added safely via addModuleStyles. This has the benefit
that it will not throw dependency errors in mobile at
the cost of the minor inconvenience of needing to add it
to the special pages that use it. On the long term rather
than using ext.uls.buttons we should use mediawiki.ui.button
or OOjs UI

Bug: T102922
Change-Id: I202f9aef491f357cefb44264093a894104692977