Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  e74184e0156be5f192811bb8e7ce51ab5fccad71
Fix mistake and documentations in CirrusSearch Hooks handlers

The code around the …?cirrusAccelerateRegex=yes debugging parameter was
introduced in Iff42e32.

Note that returning true is not necesarry in Hook handlers any more for
a long time now. Many of the "return true" can simple be removed. However,
I did not wanted to make this patch to complicated.

Change-Id: I5fabcfeace56a2cdff7daf836f38fa1ec3512be3