Updated mediawiki/extensions Project: mediawiki/extensions/PerformanceInspector  ecba8ef261a11d2eb1c395e65c04e4812930fedf
Parse the result when all promises are done

The old version only took the first result from the Promise array,
instead take them all and parse them. Also wait for all promises
to be done before generating the dialog.

This is similar to 8e321d89 which accidentally got reverted in e783081.

Change-Id: I341fac3b8dafdcbaf6fcd40f3896e8ef9416eee2