Updated mediawiki/extensions Project: mediawiki/extensions/GoogleLogin  3dce7761e53ec7a2fc7d5d29b488acd932b6ce54
Fix "Sessions are disabled for this entry point" in GoogleLogin.hooks.php

GoogleLogin currently replaces Special:UserLogin depending on if the user
is logged in and if GoogleLogin should overwrite the userlogin special page.
This results in problems, if someone (e.g. another extension) calls a function
that calls the SpecialPage_initList hook from an entry point, which doesn't
have sessions (e.g. load.php). If this happens, an exception would be thrown.

To avoid this, this change checks, if the current Request is a request for
an entry point without sessions (GoogleLogin generally doesn't need to replace
the userlogin special page for requests without sessions), and if so, doesn't
check replace the userlogin special page at all (and doesn't call the user
object). This should be considered as a hack until the extension is
converted to AuthManager (see T110294), where our own special page isn't
needed anymore (hopefully) or a better approach could be implemented.

Bug: T135445
Change-Id: I909ff5e7ef105b864516c3444ebe5bee650e3818