Updated mediawiki/extensions Project: mediawiki/extensions/WikiEditor  79a052dc5ed5cec92f24ad7d51f7dad83a68a13a
Avoid reflow due to WikiEditor toolbar loading

This avoids part of the reflow that occurs when the WikiEditor has
finished loading it's toolbar, by reserving 33px above the textarea.

I will still reflow to uncollapse a Advanced or Help section of
course, but there is only so much we can do.

Bug: T125843
Change-Id: Ibadea803bdfde9f88cff082b043ed58a05ed75a9