Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  91a882cc6f072e6d3eb28dc3310ff971da890f23
Properly filter non-core namespaces for interwiki searches

According to the InterwikiSearcher constructor we keep only core namespaces
when searching. But Searcher class does not always add the namespace filter.
It tries to optimize the query by counting the expected number of namespaces in
the target index. In the case of interwiki searches the target wiki may contain
extra content namespaces in the content index. It seems more prudent to always
add the namespace filter.
Today this is causing strange behaviors, searching Leonardo Da Vinci on itwiki
will properly display "Autore:Leonardo_da_Vinci" but the link becomes
"Progetto:Leonardo_da_Vinci" which does not exist on itwikisource.
Namespace 102 is Progetto on itwiki but Autore on itwikisource. I suppose that
core does not support specifying namespaces with interwiki links?
Anyways I find the current behaviour misleading, even if it could make sense to
search on content namespaces on the external wiki if I'm searching content
namespaces on the current wiki it seems to me more clear to explicitely always
filter on specific namespaces. Supporting this "content namespaces" behavior would
require imo to be explicitely supported in the code.

Change-Id: I1f1ca7c1a2436b4a0a2327b13950c6d03700691d