Updated mediawiki/extensions Project: mediawiki/extensions/EventBus  3517ccd69e5f32f187b41af244f24ec9e8fe68ce
Revision event URIs

Every event has a required property meta.uri. This uri is identifying
the event. Although, for most of the events except 'resource_change',
it's not 100% clear what to use as a URI, so we were using some stubs.

I propose to revisit this, and set the uris of the article-related
events to the article path, and uris of a user-related event to
the path to the user page.

Bug: T133129
Change-Id: Ie977fd73e3185ad8af0226e5d4fd8bc4d038b52e