Updated mediawiki/extensions Project: mediawiki/extensions/Echo  38adb78619685ffb1c9cc6276890b6be89a442b7
Use native MediaWiki linking to help pages

There was quite a small link under the main title in Special:Notifications
which pointed to a related help page. Now the OutputPage::addHelpLink()
method is used which moves this link to the upper right corner and allows
configuring this on-wiki (instead of LocalSettings.php or so).

Bug: T101057
Change-Id: Ib4aecee8006b8d71bb3cd86f1d4ebdfee9080870