Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseJavaScriptApi  d239f7497676c0d73958a727182444fb8adca312
Fix getLocationAgnosticMwApi behavior in Internet Explorer

* The apiEndpoint gets initialized as:
"https://www.wikidata.org/w/api.php".
That's expected. The "s" should be mentioned explicitely.

* The local counterpart is initialized as:
"//www.wikidata.org/w/api.php".
That's also expected and not wrong.

What happens is that IE returns "www.wikidata.org:443" for the first
but "www.wikidata.org" for the second. This is not the same (but should).

Bug: T136543
Change-Id: I329768d5c309f56b5eaadffd03339f2c0b03f21d