Updated mediawiki/extensions Project: mediawiki/extensions/Popups  049ae039a788341c0624565329a872a768f82bf5
Do not show Hovercards when NavPopups gadget is enabled on huwiki

This is temporary fix to get Hovercards not to show up when
Navigation Popups gadget is enabled on huwiki. The gadget module is
named 'ext.gadget.latszer' and not 'ext.gadget.Navigation_popups'.

Also check whether the global 'pg' variable is created, which means
the Navigation Popups gadget is enabled.

Bug: T135630
Change-Id: I35e1b911967200bfdfd8f44ad4e4b8dcfd844ee7