Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  d09a3eab815d4a1b55a6cc4d010d6382cfb179a8
Namespace the dns-prefetch head element item

Fix-up for Iae39a92912 (from 2012). The $name element for
OutputPage::addHeadItem() is meant to allow extensions to override head
elements of which there can be only one instance -- e.g., apple-touch-icon. But
there can be multiple <link rel=dns-prefetch> tags, and we don't want
extensions to be clobbering each other's. So use a name that is scoped to
CentralNotice, instead of the generic "dns-prefetch".

Change-Id: I2fb7f3b9f3763f2395cd3120eda5d7e2d2aaebe0