Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  7f13bed61057c0e46cff76b1243b4401980c4d4b
MT engine failure cause fallback translation not getting saved

When an MT request to cxserver fails, we fallback to source translation.
The reason for failure can be anything - very large paragraph, network
timeout, failures from the MT engines. But when the fallback to
copy-source happens, what get saved to database is the placeholder div.

This was because the section reference is not updated in the failure
handler of MT request and holding old section value. Fixed this by
moving the section reference update code to the .always() handler.

Testplan:
(if you have local cxserver, this is more easy since you can simulate
MT engine failure)
Translate Hideo Hosono from en to fr. The 'Selected publications' section
is too long for Yandex to translate and will fallback to source. You
will also see translation saved status. Reload the page by pressing F5
or CTRL+R. This section will not get restored and you will see a placeholder
with in progress indicator. Additionally you may inspect the translation
storage for the section too. With this patch, instead of placeholder,
the actual translation will get saved.

Bug: T137519
Change-Id: I72f6711a07389aa3c6b7852acd308f265aafd8a2