Updated git submodules
Project: mediawiki/extensions/CentralAuth REL1_27 fce18e7df3418721543b3982d9135a2b14e2caf1

SECURITY: Check that the loginwiki account is attached when logging in

For Special:CentralAutoLogin/checkLoggedIn, we make sure the loginwiki
user is attached. For good measure, the same is done in /validateSession
and /refreshCookies even though we shouldn't have gotten to this point
if it isn't. We don't do it for the edge wiki in /createSession or
/setCookies for reasons already explained in the comment on line 435.

For Special:CentralLogin/start, we now raise an error if the local
loginwiki account exists but isn't attached. If it doesn't exist
locally, we don't error because we assume it'll be auto-created later.

For Special:CentralLogin/complete, we check that the local edge wiki
account is attached too, even though we shouldn't have gotten to this
point if it isn't.

Bug: T137551
Change-Id: Ib2018d13ced726b6d7416fce0502d678a123f638
(cherry picked from commit 7228cdf5c43f8102e2e48391e3243f3ba6c9e3a9)