Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  4f803738e1c789763575192f7317d4a0b67a3d9b
ExtendedSearch: Show facet view when search yields no result

When search yields no result due to a selection of facets, it must
be possible to undo the last selection. This was not possible, because
the facet view was hidden when there was no result. This changed now,
so you can uncheck the facet settings that lead to a no results view.

Change-Id: Id5bc2029048d000a9854300cde1d7ea4032b219c