Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  340a5994057cf320de5b185c1a9b46b7565a7e3c
Don't strip srcset on API mobileview action

The old config hack had the effect that the mobileview API
was still using srcset, which the app relied on having for
full resolution availability (it can then at runtime strip
or use them).

Now using explicit override of the local config to restore
that behavior while being consistent with the new config.

Bug: T137608
Change-Id: Ice0e48d566e95f342c3d1554899e2715c836bc2b