Updated mediawiki/extensions Project: mediawiki/extensions/Echo  65c6b8cbaff59dbc1b151d0a78510fb5f9f53645
NotificationBadgeWidget: Fix header getting larger when "Mark all as read" present

The "Mark all as read" button had margins that made it just a little
bit taller than the header would otherwise be. This pushed the
bottom border of the header and the top border of the first notification
away from each other.

We only need top and right margins on this button, so drop the
bottom and left margins.

Bug: T137778
Change-Id: I8d47d84c79cbdad707b8cbaa8d43187c054755b3