Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  1ba410d21ce83e6fa4e4d75ea433ca7d7335da4c
Adyen iframe hacks to fix CVV input

This is version 24 of the skin, not yet live. Increases maxlength
of CVV input and adds autocomplete and inputmode hints. There
doesn't seem to be any js validation still expecting length 3.

Note that this and a lot of the CSS hiding we're doing is not
recommended. Instead, we could be sending the card type as
brandCode and sending donors to a skipDetails url.

That'll just take a bit of DI code and convincing them to give us
a card->brandCode map.

Bug: T137811
Change-Id: I46da76e1565614eddc3e45a887fe7bb324e491e9