Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  5d05f11e9b76ba8f4d4aa37e35da466e35704b8c
Allow anon users to use query API to find translation

The ApiQueryContentTranslation can be used to find a translation
by language pair and source title. It does not return the translation
content, but just the translation information - whether anybody translating
it etc. It does not make sense to restrict this for logged in users.
This commit removes that restriction.

The fetch by translation id or fetch all translations by a user demands
logging in. No change to that part.

Change-Id: Ibd538d668ddc04f22098f8a9f7174521bb2b2721