Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  5990832f89312f451e44c7efca53bf858f403b03
Checker: properly check if a page is a redirect

The Checker class only checked WikiPage::isRedirect() to determine if a page
is a redirect. Unfortunately this value can be inconsistent and is not the
one used by Update jobs.
Checker will now load and parse the page content so that it's consistent with
what the update jobs will do.
Added a --fastCheck option to the maint script to use the old behavior.

Change-Id: Iaf2d5d3c4272d21069da7947fcf4108c8e6d104e