Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  21c01a6c16a65a42bd1d21e3a7ca72552a28b3d3
Use throttled events for lazy loading images

Instead of debounced events, which seem to cause problems in browsers that
feature inertia scrolling, given it doesn't trigger as often as it needs to and
lazy images are seen unloaded when scrolling.

This should improve the UX of lazy loaded images a lot.

* Added events 'scroll:throttled' and 'resize:throttled' to mw.mobileFrontend

Bug: T137874
Change-Id: Ifeed0575de7cbcef98dc772368c7edfe790b2c36