Updated mediawiki/extensions Project: mediawiki/extensions/Echo  1d05d9298cb904b13b14e7071fc3960378dafcad
backfillUnreadWikis.php: Add rebuild mode

In rebuild mode, we only process users who already have
a row in the euw table. This reduces the number of users
we have to process significantly. Thanks to Stephane
for this idea.

Also, don't force a recache of the alert and message counts,
it's OK to fetch them from cache. We also don't need to invalidate
any caches in this script, because the transition flag from
Ibcc8ac102 will ensure that the cached global counts are
correct.

Change-Id: I69c54cac47468647df2778b2d82f1b5dd9e574d3