Updated mediawiki/extensions Project: mediawiki/extensions/Echo  92d1d600aec840f3416572bd59be0855c2ea27c5
SubGroupListWidget: use timestamp from model, not from listWidget

This ensures that the fallback timestamp is used if the list is
still empty. Not doing this causes the wikis in the cross-wiki bundle
to appear in the wrong order. Also, it makes more sense to use the
model as the source of truth regardless.

Bug: T138115
Change-Id: Icbfdc7e7c7f67179e50f0f692aef1a54568265e6