Updated mediawiki/extensions Project: mediawiki/extensions/RevisionSlider  a2d4b2cbc868f29f23eacc21de704eaa2a9a235d
Use username for username label

After changing this several times to address concerns related to
GENDER, decided to use the mediawiki standard here also used
in the core code. For translators a hint was added that GENDER
is not supported here so an appropriate translation should be
chosen by them.

Bug: T136367
Change-Id: Icc2a8e6fc08761367c4ab6bb2a446262f8aa86e6