Updated mediawiki/extensions Project: mediawiki/extensions/AbuseFilter  6af0857c553833cde9cf600769dcabcda2aa2b67
Move the filter pre-caching outside of the DB lock

Edits should not wait on this to finish, especially since it does
both the minor and non-minor edit cases, only one of which actually
being useful. If the cache is there in time and there is no edit
summary it will be used; if not, there probably no reason to wait.

Bug: T138550
Change-Id: Ifc3b97ddf4dbb94f8ec3eacfcd5c8994c69aafbc