Updated mediawiki/extensions Project: mediawiki/extensions/Echo  e15c258d7b6b8e16ac1e89450b8783870e5f210e
Remove initiallyUnseen before marking item read/unread

If we are marking the item as read/unread, we don't want the change
in class to trigger the unseen animation. When that happens, first
remove the 'initiallyUnseen' class (that triggers the animation)
and only then mark the item as read or unread.

Bug: T138522
Change-Id: I05e02518feb1fc166297594053a084718dcb7194