Updated mediawiki/extensions Project: mediawiki/extensions/Kartographer  b5b92fa47da6b99c19a21cb275051a9a51c313ea
Fix auto position/zoom not working on maplink.

There was an issue with data being loaded asynchronously
and the rest of the code remaining being executed
synchronously. This refactor introduces a utility "ready"
helper that gets the map ready and execute a handler
with the map object and the updated map data object.

The code that must be executed when the map is "ready"
should use this utility method and add a handler.

Bug: T138797
Change-Id: Ibe2c1a5988022fa9eb1ec4e021087731ca8cd1c0