Updated mediawiki/extensions Project: mediawiki/extensions/RevisionSlider  6101897460dd0827775439dee9ee1942ed3d2934
Add browser tests for tutorial / help dialog

Testing the visibility / presence of things around
the dialog appears to be hard and the DOM for the dialog
remains unchanged when shown / hidden as far as I can see.

These tests test that the dialog is only shown once per
user.
And that the sequence of the dialog works in regards to the correct
buttons being present at the correct times.

Bug: T133278
Change-Id: Ia86cb69573da3e888c2897b8b50f1d2e5c61d8d4