Updated mediawiki/extensions Project: mediawiki/extensions/CollaborationKit  cb7e05ed9d3cad8bad8c4028a2159eade1f27ef1
Make normal link when item is a category or file.

Change-Id: I0a7cd12589c274761e9b329068638d5f0b2c96cc

Add a human readable edit format.

Not 100% sure this is the best approach.

Possibly we want to pop the description out into a separate
text box. Unclear how easy that is to do with this approach,
possibly we can use EditPage hooks, but EditPage is kind of
scary.

Ideally we would have a different default format for user editing than
we would for API editing/db storage, but that doesn't seem possible.

Also diffs seem to be broken, but that seems to be a core MW bug.

Bug: T139097
Change-Id: I3035889e993e9dd020ac8b09fae71fbecfbd6e77