Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  651ac786d760595f5799d43ce9050ddfa9d7f4c2
Consistently pass api in uw.controller.Step subclasses

I started doing this because I needed a mw.Api instance in
uw.controller.Upload, and kind of became unsure in the middle
if this really makes the code better. But it definitely
doesn't make it worse, and it is more consistent I guess.
I updated and synchronized the documentation comments too.

Change-Id: Ib894887a5214efb0774f56cd32688cf1af8d2d5b