Updated mediawiki/extensions Project: mediawiki/extensions/AbuseFilter  8ee834c376baddce4793f549ff008b6f045c7ccd
AFPUserVisibleException should log in English, not the user's language

The user's language should only be used when the exception is actually
displayed to the user.

This will also avoid "User::loadFromSession called before the end of
Setup.php" warnings when the syntax error is encountered during filter
execution for account autocreation, where we don't display it to the
user.

Bug: T124367
Change-Id: Ic17f56aecbe575ef15c6970c4298f889249e1904