Updated git submodules
Project: mediawiki/extensions/Wikispeech master 227e1fe067f041b06927c010485a011945eafe7b

Start and stop recitation with mouse

Added functionality for starting and stopping recitation by clicking a button
on the page. This is prepared by adding utterance tags to the page, containing
the utterance as a string and an audio tag to use for playing the resulting
audio. When an uttrance is played, the utterance string is sent to the TTS
server. The URL for the resulting audio file is retrieved from the response
and set as source for the audio element.

Updated PHP code to run with PHP 5.5.

Bug: T133672
Change-Id: Ie040eaa780576f16ed4cb1c1be6d223c99540fc1