Updated git submodules
Project: mediawiki/extensions/Translate master c7c9ee42abc001d5a3b86700c324c5fa2ca686da

PageTranslation: Do renders for manually fuzzied units as well

In page translation, if a unit gets fuzzied automatically, the normal
behavior is to update the translation unit with the text from the source
unit. So for consistency, even if a unit is marked with !!FUZZY!! manually,
render the translation with the source unit's text instead of not doing an
update at all.

Bug: T137715
Change-Id: I69203dc33ec2eed88f9320e421c4e59affa713c3