Updated git submodules
Project: mediawiki/extensions/Wikibase master bf44f7529c4fb6d6779d0c4268b9cd120fcd8561

Don't check the namespace in SidebarHookHandlers::doSidebarBeforeOutput

Why would the "wikibase_item" property be set, if we're not in
a Wikibase namespace. Needed to make this also work for article
placeholders which are not in a wikibase namespace, but act
similar to pages in one.

Change-Id: I8288a625901889f29ad889d6c8f1f636b9a3453d