Updated git submodules
Project: mediawiki/extensions/CentralNotice master 58fd6c3135387538fd17eae1ee27240b575ba679

GeoIP: Background lookup optional and configurable, provide data via promise

Removes hard-coded fallback background call to geoiplookup.wikimedia.org. If a
background call to a third-party service is desired when no GeoIP cookie is
present, it may be configured using the new global
$wgCentralNoticeGeoIPBackgroundLookupModule.

An example module to call freegeoip.net is included, but is not activated in
the default config.

Also, resolves the geoIP promise with geo data, and adapts CentralNotice to
use that instead of the global window.Geo. This is a first step towards
possibly deprecating that global.

Bug: T143271
Change-Id: I4b7686eb7935c802a3b2b1d36c21dc0521f8e9a6