Updated git submodules
Project: mediawiki/extensions/UploadWizard master f35c67946392a14b240eac56f4e62841695517b6

Don't display place for tutorial if it is skipped

This only works if the global or campaign settings say to skip it.  User
prefs that skip it still show the place.

This makes sense because the user cannot use the back button to reach
the “Learn” step, so showing that the step is being skipped is confusing.

Bug: T144963
Change-Id: I126a38c8bad4077e0dc885169f9be5ff55778577