Updated git submodules
Project: mediawiki/extensions/MediaWikiFarm master a5ca6b205f178d0de16982011a8811d6daf4f38f

Mechanism to fallback to cached version in case of syntax error

If a syntax error is introduced in a config file, previously it
was fatal until the error is solved. With this mechanism and when
cache is activated, it fallbacks to cached version in case of
unreadable file (now-missing file or bad-syntax file).

A new object property $errors is introduced; for now it has only
this type of error, but I will probably create catories of errors
in the future. A mechnism to output there errors somewhere is a
next step, but I have no exact plans for now.

Change-Id: Idf6aa8cbcdcc93d2c9fb890c30f30891266e2c73