Updated git submodules
Project: mediawiki/extensions/MobileFrontend master 1ac2bf5688fce58ce46ba20587480f3ff6fd3e09

Get height and width from style to placeholder

This reverses the mistake in T143768 and instead copies the dimensions
from the style attribute to the element so that a placeholder element
shouldn't be without dimensions.

It is possible that there are elements without width and height
attributes or styles but we ignore this case.

Other styles are not passed to the image as due to the different HTML
structure they could lead to unexpected side effects as we discovered
in I0a42887429fa64034a04bdb53eae9b8814cd1570.

Bug: T145222
Change-Id: If2b81135d52601614803c4834dd15bde7a77aea9