Updated git submodules
Project: mediawiki/extensions/Babel master cd12336cb0a5f7c1f7c64495e2a864c3e2ffdae2

Store babel languages in the database

The current method of reading categories and parsing languages out of
them is a terrible hack, and a TODO comment suggests to use a database,
which is a much better structured storage format.

This adds a `babel` database table, which if $wgBabelUseDatabase is
true, will be used to store the languages a user knows, and at what
level.

Once the database is populated and created everywhere, we should remove
the fallback category code.

Change-Id: I520358ac778cfddb1cf768c798f322dce61922fd