Updated git submodules
Project: mediawiki/extensions/Flow master c23d4e637dfff8b027ca4ff0df47926ca396599c

Don't crash when a revision is not found

In some cases (to be discovered), the recent changes
or watchlist page crashes with "Revision not found in revisionCache".

This change logs a warning and returns false for that specific
revision instead of throwing an exception. It's only a bandaid
until we fix the problem and will allow us to collect more
information if we're not able to fix it now.

Bug: T145125
Change-Id: I74de54336aec8f9c6d709d60ca4c6f655bb5d46d