Updated git submodules
Project: mediawiki/extensions/Popups master 9f92324b4d05bdd8555048cc53eea9f160bc773f

Dismiss event should not be tied to render method

The dismiss event is currently a side effect of closing a popup.
This should be more closely coupled with the event that causes it.

This allows the closePopup method to be called without trigging
an event which is a necessary precursor to closing popups when
navigating away from the page.

Change-Id: I3f3c4780cc31dc8d84cdd76df2c77fa45fbea882