Updated git submodules
Project: mediawiki/extensions/MobileFrontend master 558ee951732b27716b73cfcdece8d16f40afd4a9

Add messages used in sources to the modules where they are used

When a file references a message, then in the extension.json
configuration it has to reference it in the `messages` key.

Fix some instances of this messages not being there.

Bug: T146748
Change-Id: If2d777dabcce1b1d140098bd3622a2e409bd512e