Updated git submodules
Project: mediawiki/extensions/CollaborationKit master ed9cc90097bd2bdcb4c1a7c5b260515f612e4df3

Add AllowedNamespaces configuration variables for hubs and lists

Users* may want to change the namespaces in which hubs/lists are
allowed in order to fullfill unexpected use cases. We don't care;
we'll just leave it an option.

Separate variables are used for hubs and lists as uses may vary for
each.

Hubs now also specifically check if subpages are allowed for the
namespace in case a user misconfigures their wiki.

Allowed namespaces are set to user (2) and project (4) by default.
Unless that's the other way around. I've forgotten already, but it's
those two.

* All mentions of 'users' here refers to sysadmins with localsettings
access, not onwiki users.

This is a follow-up on initial implementation for T141004.

Change-Id: If70bff87c8324f425d32511d94beaac17017cf41