Updated git submodules
Project: mediawiki/extensions/ProofreadPage master edf1a188b83248c9c1d98a5924888f6dc8f902a3

Merge "Reduce main edit box height when editin in vertical mode"

Reduce main edit box height when editin in vertical mode

This change makes the main text box smaller so that it all can be
visible on the screen along with the image above it. Both are set
to the same height of one-third of the window height.

When switching back to the default horizonal mode, the image and
the text box are returned to their default sizes.

This also fixes a bug whereby the zoom initialization was setting
the image height too late, and so the image was displaying at full
height.

Bug: T145923
Change-Id: I6d2ec037f0633c413e3d225c2477e29f73200b36