Updated git submodules
Project: mediawiki/extensions/CollaborationKit master 1213fc752b1aaa94f9cf44d30fe1e0ff9a0f67cd

When a CollaborationListContent or CollaborationHubContent page does not meet the schema, instead of throwing an exception it returns the offending JSON.

This introduces a new top-level property in the object called "displaymode" which is currently optional but will be required soon. The options are "normal" (the default), "error" (used for outputting the error) and "members" (CollaborationListContent only; not currently used but will replace the ismemberlist bool).

Bug: T148085
Change-Id: I4f76f8f43405cf81eb84108a848a76e0255a76b7