Updated git submodules
Project: mediawiki/extensions/TwoColConflict master da5bbaaa04a26cc1aa019dc80d5036a305e96dd5

Replace the conflict page with a two column version

This is the first version of the two column edit conflict page. A new
class is introduced extending the EditPage to show the two column version.

Basically this version moves the editor with the conflicting change from
the bottom to the top into a column next to the editor with the version
that will be saved when the conflict is solved.

Next steps will be addressed in follow-up patches.

Most parts of the code if the original EditPage are still used to avoid
interferences with other bits of the edit page logic.

Bug: T143823
Change-Id: Ibfb2bfafa4f5837ef3a02407625c0f579527095a
Depends-On: I98df04f5445dcc87a96cde157df13786414b8c3a